move index.js to example folder
