+++ example/data/ColorScheme.json
